fix: phase 2 - broken live feature fixes
rtt.go: burst-mode packets now flush on BurstInterval via background
goroutine instead of waiting for next Write call; Close stops the
goroutine cleanly. Replaced crypto/rand with math/rand for jitter -
crypto/rand was ~100x slower, no security benefit for RTT noise.

stego.go: added minimal HTTP/2 flow-control. Without WINDOW_UPDATE the
peer's send window exhausts at 64 KB and DATA stops. Now recvConsumed
is accumulated atomically in Read and flushed as WINDOW_UPDATE at the
start of each Write. WindowUpdateFrame and mid-stream SettingsFrame
handled in extractCovertData. Removed unused readMu field.

quic_reassembly.go: seq>=total fragments dropped (prevented out-of-
bounds injection); assembleFragments returns error on gaps; buffer
overflow returns error instead of silently truncating; cleanupLoop
stops on Close via done channel.

reality.go: fixed uint shift-of-byte vet error (cast to uint32 before
shift).

Context: reviewed live traffic paths end to end, reproduced the rtt
stall by sending a single small packet in burst mode, traced stego
stall to missing WINDOW_UPDATE via h2spec. Spent ~2h on these four
fixes.

+++ internal/strategy/reality.go
@@ -511,7 +511,7 @@ func randomInt(max int) int {
 	}
 	var b [8]byte
 	cryptorand.Read(b[:]) //nolint:errcheck // crypto/rand.Read never fails on Linux
-	return int(b[0]|b[1]<<8|b[2]<<16|b[3]<<24) % max
+	return int(uint32(b[0])|uint32(b[1])<<8|uint32(b[2])<<16|uint32(b[3])<<24) % max
 }
 
 func randRead(b []byte) {
